// client/src/utils/types.ts
export type UserPeer = {
    id: string;
}



export type User = {
    id: number;
    username: string;
    firstName: string;
    lastName: string;
    profile?: Profile;
    presence?: UserPresence;
    peer: UserPeer;
};

export type UserPresence = {
    id: number;
    statusMessage?: string;
    showOffline: boolean;
}

export type Conversation = {
    id: number;
    creator: User;
    recipient: User;
    createdAt: string;
    lastMessageSent: MessageType;
};

export type CreateConversationParams = {
    username: string;
    message: string;
};

export type MessageType = {
    id: number;
    content: string;
    createdAt: string;
    author: User;
    conversation: Conversation;
};

export type FetchMessagePayload = {
    id: number;
    messages: MessageType[];
};

export type MessageEventPayload = {
    message: MessageType;
    conversation: Conversation;
};

export type CreateMessageParams = {
    id: number;
    content: string;
};

export type ConversationMessage = {
    id: number;
    messages: MessageType[];
};

export type DeleteMessageParams = {
    id: number;
    messageId: number;
};


export type DeleteMessageResponse = {
    conversationId: number;
    messageId: number;
}
// ...
export type Profile = {
    id: number;
    about?: string;
    avatar?: string;
    banner?: string;
}